refactor(cli): extract status-bar/skill-command mixins from cli.py (shard s2)
